+++ view/packages/wikibase-data-values-value-view/src/experts/QuantityInput.js
@@ -54,15 +54,10 @@
 		/**
 		 * @inheritdoc
 		 */
-		valueCharacteristics: function( format ) {
+		valueCharacteristics: function() {
 			var options = {
 				unit: this._unitSelector && this._unitSelector.getConceptUri() || null
 			};
-
-			if ( format === 'text/plain' ) {
-				options.applyRounding = false;
-				options.applyUnit = false;
-			}
 
 			return options;
 		},