+++ view/packages/wikibase-data-values-value-view/tests/src/resources.php
@@ -39,7 +39,6 @@ return call_user_func( function() {
 				'jquery.valueview.valueview',
 				'test.sinonjs',
 				'valueFormatters.formatters',
-				'valueFormatters.ValueFormatterStore',
 				'valueParsers.parsers',
 				'valueParsers.ValueParserStore'
 			),