+++ view/packages/wikibase-data-values-value-view/RELEASE-NOTES.md
@@ -1,8 +1,13 @@
 # ValueView release notes
 
+## 0.16.0 (dev)
+
+* Changed `ValueView` to take two `ValueFormatter` instances instead of a
+  `ValueFormatterStore`
+
 ## 0.15.9 (2016-02-03)
 
-* Correctly detect a changed language code when editing a `MonolingualTextValue`
+* Correctly detect a changed language code when editing a `MonolingualTextValue`.
 
 ## 0.15.8 (2016-01-27)